rename, amend paths
