+++ app/lib/alimonitor-services/MonalisaService.js
@@ -31,7 +31,7 @@ const { databaseManager: {
     },
     sequelize,
 } } = require('../database/DatabaseManager.js');
-const { findOrCreatePeriod } = require('../services/periods/findOrUpsertOrCreatePeriod.js');
+const { findOrCreatePeriod } = require('../services/periods/findOrUpdateOrCreatePeriod.js');
 
 class MonalisaService extends AbstractServiceSynchronizer {
     constructor() {
